requirements.txt
anyio==2.2.0
argon2-cffi==20.1.0
astor==0.8.1
asttokens==2.0.5
async-generator==1.10
attrs==20.3.0
autograd==1.3
autograd-gamma==0.5.0
Babel==2.9.0
backcall==0.2.0
black==21.12b0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.3
executing==0.8.2
fonttools==4.28.5
formulaic==0.2.3
future==0.18.2
idna==2.10
interface-meta==1.2.3
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.1
jupyter-server==1.6.1
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
kiwisolver==1.3.1
lark-parser==0.11.3
lifelines==0.25.11
MarkupSafe==1.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.4.1
portion==2.1.6
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.1
python-intervals==1.10.0
pytz==2021.1
pywin32==303
pywinpty==1.1.6
pyzmq==22.0.3
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.1.4
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tomli==1.2.3
tomlkit==0.7.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing_extensions==4.0.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
